indentation++
svn path=/trunk/KDE/kdenetwork/kget/; revision=833189

+++ transfer-plugins/bittorrent/bttransfer.cpp
@@ -261,7 +261,7 @@ void BTTransfer::updateTorrent()
         changesFlags |= Tc_DownloadSpeed;
     }
 
-    if(m_percent != (m_percent = ((float) chunksDownloaded() / (float) chunksTotal()) * 100) )
+    if (m_percent != (m_percent = (float)(chunksDownloaded() / chunksTotal()) * 100))
         changesFlags |= Tc_Percent;
 
     setTransferChange(changesFlags, true);
